알람 설정 안되는 오류 수정
