+++ app/src/main/java/com/donghaeng/withme/screen/SplashActivity.java
@@ -7,6 +7,7 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
+import android.provider.AlarmClock;
 import android.provider.Settings;
 import android.util.Log;
 import android.widget.Toast;
@@ -35,6 +36,7 @@ public class SplashActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_splash);
+        handleIntent(getIntent());
 
         // 권한 체크 시작
         checkPermissions();
@@ -168,4 +170,38 @@ public class SplashActivity extends AppCompatActivity {
                     ScreenChecker.getInitialScreen(this);
                 });
     }
+
+    // 알람 설정 관련 메소드 들
+    @Override
+    protected void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        handleIntent(intent);
+    }
+    private void handleIntent(Intent intent) {
+        if (intent != null && intent.getBooleanExtra("SET_ALARM", false)) {
+            int hour = intent.getIntExtra("ALARM_HOUR", -1);
+            int minute = intent.getIntExtra("ALARM_MINUTE", -1);
+
+            if (hour != -1 && minute != -1) {
+                setAlarm(hour, minute);
+            }
+        }
+    }
+    private void setAlarm(int hour, int minute) {
+        Intent alarmIntent = new Intent(AlarmClock.ACTION_SET_ALARM)
+                .putExtra(AlarmClock.EXTRA_HOUR, hour)
+                .putExtra(AlarmClock.EXTRA_MINUTES, minute)
+                .putExtra(AlarmClock.EXTRA_MESSAGE, "알람")
+                .putExtra(AlarmClock.EXTRA_SKIP_UI, true);
+
+        try {
+            startActivity(alarmIntent);
+        } catch (Exception e) {
+            Toast.makeText(this,
+                    "알람 설정 중 오류가 발생했습니다: " + e.getMessage(),
+                    Toast.LENGTH_LONG
+            ).show();
+        }
+    }
+
 }
